Add curl-free correction for the magnetic field
The correction is taken from [1].  It is a 2nd order correction to make
the magnetic field consistent with experimentally available fields,
i.e., to make curl-free and divergence free (which it already was).  The
correction works for 1D axially varying fits of radial fields and
produces a 2D field with components in the axial direction and in the
same direction of the field, i.e,
  (0, B_y(z), 0) -> (0, B_y(z)*, B_z(z)).

A more general approach need be found to account for more general
magnetic fields, e.g., non-normal non-uniform fields.

The implementation of this correction makes the `dirB` parameter useless
for the non-`:uniform` B fields, as in those cases the field direction
is fixed per definition of the correction.

[1]: X. Albets-Chico et al. (2011), Fusion Eng. Des. 86(1), 5-14.